+++ test/e2e/features/upgrade/suite.go
@@ -4,16 +4,24 @@ package upgrade
 
 import (
 	"context"
+	"fmt"
 	"net/http"
+	"os"
 	"path/filepath"
+	"strconv"
 	"strings"
 
+	"github.com/google/go-github/v67/github"
+	"github.com/shurcooL/githubv4"
 	"github.com/stretchr/testify/suite"
+	"golang.org/x/oauth2"
 	"k8s.io/apimachinery/pkg/types"
 
+	"github.com/kgateway-dev/kgateway/v2/pkg/utils/cmdutils"
 	"github.com/kgateway-dev/kgateway/v2/pkg/utils/fsutils"
 	"github.com/kgateway-dev/kgateway/v2/pkg/utils/kubeutils/kubectl"
 	"github.com/kgateway-dev/kgateway/v2/pkg/utils/requestutils/curl"
+	"github.com/kgateway-dev/kgateway/v2/pkg/utils/threadsafe"
 	"github.com/kgateway-dev/kgateway/v2/pkg/version"
 	"github.com/kgateway-dev/kgateway/v2/test/e2e"
 	"github.com/kgateway-dev/kgateway/v2/test/e2e/common"
@@ -25,8 +33,9 @@ import (
 )
 
 var (
-	_             e2e.NewSuiteFunc = NewTestingSuite
-	setupManifest                  = filepath.Join(fsutils.MustGetThisDir(), "testdata", "setup.yaml")
+	_                        e2e.NewSuiteFunc = NewTestingSuite
+	setupManifest                             = filepath.Join(fsutils.MustGetThisDir(), "testdata", "setup.yaml")
+	GatewayNameLabelSelector                  = "gateway.networking.k8s.io/gateway-name=gateway"
 )
 
 // testingSuite validates that kgateway can be upgraded from a released version to the locally-built chart.
@@ -35,11 +44,23 @@ var (
 //   - Uninstalling kgateway after this suite completes.
 type testingSuite struct {
 	*base.BaseTestingSuite
+	podSelectors []string
+}
+
+func NewTestingSuiteWithConfig(ctx context.Context, testInst *e2e.TestInstallation, podSelectors []string) func(ctx context.Context, testInst *e2e.TestInstallation) suite.TestingSuite {
+	return func(ctx context.Context, testInst *e2e.TestInstallation) suite.TestingSuite {
+		return &testingSuite{
+			BaseTestingSuite: base.NewBaseTestingSuite(ctx, testInst, base.TestCase{}, nil),
+			podSelectors:     podSelectors,
+		}
+	}
 }
 
 func NewTestingSuite(ctx context.Context, testInst *e2e.TestInstallation) suite.TestingSuite {
 	return &testingSuite{
-		base.NewBaseTestingSuite(ctx, testInst, base.TestCase{}, nil),
+		BaseTestingSuite: base.NewBaseTestingSuite(ctx, testInst, base.TestCase{}, nil),
+		// This should be updated to app.kubernetes.io/component=proxy. v2.2.x did not have this label
+		podSelectors: []string{GatewayNameLabelSelector},
 	}
 }
 
@@ -86,8 +107,9 @@ func (s *testingSuite) TestUpgrade() {
 	s.assertNoKgatewayPodErrors()
 
 	// Ensure the proxy pod is also updated
-	// This should be updated to app.kubernetes.io/component=proxy. v2.2.x did not have this label
-	s.TestInstallation.AssertionsT(s.T()).EventuallyPodHasImageVersion(s.Ctx, "default", "gateway.networking.k8s.io/gateway-name=gateway", version.Version)
+	for _, selector := range s.podSelectors {
+		s.TestInstallation.AssertionsT(s.T()).EventuallyPodHasImageVersion(s.Ctx, "default", selector, version.Version)
+	}
 
 	// Ensure the same gateway works after the upgrade
 	common.BaseGateway.Send(
@@ -130,4 +152,111 @@ func (s *testingSuite) assertNoKgatewayPodErrors() {
 			)
 		}
 	}
+}
+
+func newGraphQLClient() *githubv4.Client {
+	token := os.Getenv("GITHUB_TOKEN")
+	if token == "" {
+		return githubv4.NewClient(http.DefaultClient)
+	}
+	ts := oauth2.StaticTokenSource(&oauth2.Token{AccessToken: token})
+	return githubv4.NewClient(oauth2.NewClient(context.Background(), ts))
+}
+
+// fetchGithubRelease pages through kgateway releases ordered by creation date descending
+// (newest first, skipping drafts) using the GitHub GraphQL API, which
+// guarantees the ordering via orderBy. Returns the tag name of the first release for
+// which match returns true.
+func fetchGithubRelease(ctx context.Context, match func(tagName string) (bool, error)) (string, error) {
+	client := newGraphQLClient()
+
+	var query struct {
+		Repository struct {
+			Releases struct {
+				Nodes []struct {
+					TagName      githubv4.String
+					IsDraft      githubv4.Boolean
+					IsPrerelease githubv4.Boolean
+				}
+				PageInfo struct {
+					EndCursor   githubv4.String
+					HasNextPage githubv4.Boolean
+				}
+			} `graphql:"releases(first: 100, after: $cursor, orderBy: {field: CREATED_AT, direction: DESC})"`
+		} `graphql:"repository(owner: \"kgateway-dev\", name: \"kgateway\")"`
+	}
+
+	variables := map[string]any{
+		"cursor": (*githubv4.String)(nil),
+	}
+
+	for {
+		if err := client.Query(ctx, &query, variables); err != nil {
+			return "", fmt.Errorf("graphql query releases: %w", err)
+		}
+		for _, r := range query.Repository.Releases.Nodes {
+			if bool(r.IsDraft) || string(r.TagName) == "" {
+				continue
+			}
+			ok, err := match(string(r.TagName))
+			if err != nil {
+				return "", err
+			}
+			if ok {
+				return string(r.TagName), nil
+			}
+		}
+		if !bool(query.Repository.Releases.PageInfo.HasNextPage) {
+			break
+		}
+		variables["cursor"] = githubv4.NewString(query.Repository.Releases.PageInfo.EndCursor)
+	}
+	return "", fmt.Errorf("no matching release found")
+}
+
+// FetchLatestRelease returns the most recent release tag that is an ancestor of HEAD.
+// This mirrors `git describe --tags --abbrev=0` but works in shallow checkouts where
+// tags are not fetched, by resolving HEAD via git then checking ancestry via the GitHub API.
+func FetchLatestRelease(ctx context.Context) (string, error) {
+	var shaOut threadsafe.Buffer
+	if err := cmdutils.Command(ctx, "git", "rev-parse", "HEAD").
+		WithStdout(&shaOut).
+		WithStderr(&shaOut).
+		Run(); err != nil {
+		return "", fmt.Errorf("git rev-parse HEAD: %w", err)
+	}
+	headSHA := strings.TrimSpace(shaOut.String())
+
+	// Use REST client for CompareCommits — no GraphQL equivalent.
+	restClient := github.NewClient(nil)
+	if token := os.Getenv("GITHUB_TOKEN"); token != "" {
+		restClient = github.NewClient(nil).WithAuthToken(token)
+	}
+
+	return fetchGithubRelease(ctx, func(tagName string) (bool, error) {
+		// Compare tag...HEAD: status=="ahead" means HEAD is ahead of the tag (tag is an ancestor).
+		comparison, _, err := restClient.Repositories.CompareCommits(ctx, "kgateway-dev", "kgateway", tagName, headSHA, nil)
+		if err != nil {
+			return false, fmt.Errorf("compare %s...%s: %w", tagName, headSHA, err)
+		}
+		return comparison.GetStatus() == "ahead" || comparison.GetStatus() == "identical", nil
+	})
+}
+
+func FetchPreviousMinorRelease(ctx context.Context, latestRelease string) (string, error) {
+	parts := strings.Split(latestRelease, ".")
+	if len(parts) < 2 {
+		return "", fmt.Errorf("unexpected tag format: %s", latestRelease)
+	}
+	minorInt, convErr := strconv.Atoi(parts[1])
+	if convErr != nil {
+		return "", fmt.Errorf("failed to parse minor version from tag %q: %v", latestRelease, convErr)
+	}
+	if minorInt <= 0 {
+		return "", fmt.Errorf("no previous minor for release %q (minor is 0)", latestRelease)
+	}
+	previousMinorPrefix := fmt.Sprintf("%s.%d.", parts[0], minorInt-1)
+	return fetchGithubRelease(ctx, func(tagName string) (bool, error) {
+		return strings.HasPrefix(tagName, previousMinorPrefix), nil
+	})
 }
